// payload-types.ts
/**
 * This file was generated by Payload.
 * It is provided for reference only and should not be modified.
 */

export interface Config {
  collections: {
    users: User
    profile: Profile
    experience: Experience
    projects: Project
    skills: Skill
    media: Media
  }
  globals: {}
}

export interface User {
  id: string
  name?: string | null
  email: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  password: string | null
  createdAt: string
  updatedAt: string
}

export interface Profile {
  id: string
  name: string
  title: string
  bio: any
  email: string
  phone: string
  location: string
  resumeURL: string | Media
  socialLinks?: {
    platform: 'linkedin' | 'github' | 'twitter' | 'email'
    url: string
    id?: string | null
  }[] | null
  seoTitle: string
  seoDescription: string
  seoKeywords?: {
    keyword?: string | null
    id?: string | null
  }[] | null
  createdAt: string
  updatedAt: string
}

export interface Experience {
  id: string
  company: string
  role: string
  startDate: string
  endDate?: string | null
  description: any
  techStack: {
    technology?: string | null
    id?: string | null
  }[]
  logo?: string | Media | null
  order: number
  createdAt: string
  updatedAt: string
}

export interface Project {
  id: string
  title: string
  category: string
  description: any
  techStack: {
    technology?: string | null
    id?: string | null
  }[]
  repoLink?: string | null
  liveLink?: string | null
  coverImage: string | Media
  featured?: boolean | null
  order: number
  slug: string
  createdAt: string
  updatedAt: string
}

export interface Skill {
  id: string
  name: string
  iconName: string
  category: 'frontend' | 'backend' | 'database' | 'devops' | 'tools'
  proficiency?: ('1' | '2' | '3' | '4' | '5') | null
  order: number
  createdAt: string
  updatedAt: string
}

export interface Media {
  id: string
  alt: string
  createdAt: string
  updatedAt: string
  url?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
}